Link Vim's `SpecialKey` to `SrceryXgray4`
Some characters in `listchars` are displayed using the `SpecialKey`
highlight group. For the Srcery theme, this is linked to `SrceryBlue`,
which is fine for displaying things like `<Enter>` or `<F1>`, but it
also affects the `leadmultispace` and `tab` characters from `listchars`.

Linking `SpecialKey` to `SrceryXgray4` would mean that it'd also affect
things like `<Enter>` or `<F1>` etc. But I guess I can live with that
for now.