Remove Vim Test package
The bindings added for the package (<leader>tn) was clashing with the
binding to open a new tab. The initial thought was to remove or change
the binding itself, but then I realised that the Vim Test package isn't
required as I use the native compiler feature of Vim now.