Add gtags.conf
This is a heavily modified version of the file that comes with the GNU
Global program. Basically, I've removed everything that I don't deal
with, and made the Pygments parser the default.

Pygments uses the ctags program for some reason (probably for generating
the tag files), so one thing that's not taken care of is to ensure that
gtags is able to find this parser on both Linux and Mac environments.
Currently it's hardcoded that the location of the ctags executable is at
/usr/bin/ctags. Which means it probably won't work on Mac. And also
probably might not work on other Linux environments except the one used
to create this config!